requirements.txt
google-generativeai
python-dotenv
streamlit
pypdf
